Integrate BibDetails Functional With Original BibDetails
Inorder to retain the the diff between the files it was asked to integrate the two files.
