+++ src/app/components/BibPage/BottomBibDetails.jsx
@@ -2,7 +2,7 @@ import { Heading } from '@nypl/design-system-react-components';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { definitionMarcs } from '../../utils/bibDetailsUtils';
-import BibDetails_Functional from './BibDetails_Functional';
+import BibDetails from './BibDetails';
 
 // `linkable` means that those values are links inside the app.
 // `selfLinkable` means that those values are external links and should be self-linked,
@@ -49,7 +49,7 @@ const BottomBibDetails = ({ bib }) => {
   return (
     <section style={{ marginTop: '20px' }}>
       <Heading level={3}>Details</Heading>
-      <BibDetails_Functional fields={fields} marcs={definitionMarcs(bib)} />
+      <BibDetails fields={fields} marcs={definitionMarcs(bib)} />
     </section>
   );
 };
